move modules to redux folder
