+++ src/PrivateRoute.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Route, Redirect } from "react-router-dom";
 import { connect } from "react-redux";
-import { userIsLoggedIn } from "./modules/user";
+import { userIsLoggedIn } from "./redux/modules/user";
 
 export const PrivateRoute = connect(() => ({
   isLoggedIn: userIsLoggedIn,
