+++ src/redux/modules/map/actions.js
