+++ src/redux/modules/map/reducer.js
