+++ src/redux/modules/map/selectors.js
