+++ src/redux/modules/user/reducer.spec.js
